Add "Other Meetup Groups" layer
This additional layer uses a triangle icon based on the HTC pyramids,
as well as the HTC pyramid color. Changed the blue circle color to use
the HTC pyramid color as well. Also, changed circle from fixed area to
marker with fixed size in pixels so as to be consistent with triangle
icons.